% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sPSE_true value.R
\name{iPSEsv_t}
\alias{iPSEsv_t}
\title{General approach of causal mediation analysis for survival outcome under sequential mediators.}
\usage{
iPSEsv_t(mediators, alpha_Y, beta_Y, gamma_Y, delta_Y, alpha_M, beta_M, gamma_M,
  delta_M, sigma2_M, alpha_C, beta_C, gamma_C, delta_C, sigma2_C,
  C_hat = c(0.5), a1 = 1, a0 = 0)
}
\arguments{
\item{mediators}{The name of mediators.}

\item{parameters}{The real values of parameters}
}
\value{
A list of true iPSEsv and partPSEsv
}
\description{
Derive the true iPSEsv based on real parameters.
}
\examples{
TrueV <- iPSEsv_t(mediators=c("x1","x2"),
alpha_Y=matrix(1,1,1),beta_Y=matrix(1),gamma_Y=matrix(c(0.5,0.6),1,2),
delta_Y=matrix(c(0.5,0.6),1,2),
alpha_M=matrix(1,2,2),beta_M=matrix(1,2,1),gamma_M=matrix(1,2,2),
delta_M=matrix(1,2,1),sigma2_M=matrix(1,2,1),
alpha_C=matrix(1,2,2),beta_C=matrix(1,2,1),gamma_C=matrix(1,2,2),
delta_C=matrix(1,2,1),sigma2_C=matrix(1,2,1),
C_hat = c(1,0.5),
a1=1,a0=0)
}
\author{
An-Shun Tai \email{daansh13@gmail.com}, Pei-Hsuan Lin \email{a52012232@gmail.com}, and Sheng-Hsuan Lin \email{shenglin@nctu.edu.tw}
}
